HSEARCH-3349 Move VerifyingStubBackendBehavior to its own file
It's getting too big to just be an inner class.
