+++ util/internal/integrationtest/common/src/main/java/org/hibernate/search/util/impl/integrationtest/common/rule/BackendMock.java
@@ -9,20 +9,14 @@ package org.hibernate.search.util.impl.integrationtest.common.rule;
 import java.lang.invoke.MethodHandles;
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.Objects;
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
 
 import org.hibernate.search.engine.backend.work.execution.DocumentCommitStrategy;
 import org.hibernate.search.engine.backend.work.execution.DocumentRefreshStrategy;
-import org.hibernate.search.engine.backend.types.converter.runtime.FromDocumentFieldValueConvertContext;
 import org.hibernate.search.engine.search.DocumentReference;
-import org.hibernate.search.engine.search.loading.context.spi.LoadingContext;
-import org.hibernate.search.engine.search.query.SearchResult;
 import org.hibernate.search.util.common.logging.impl.Log;
 import org.hibernate.search.util.common.logging.impl.LoggerFactory;
 import org.hibernate.search.util.impl.integrationtest.common.stub.backend.StubBackendBehavior;
@@ -31,7 +25,6 @@ import org.hibernate.search.util.impl.integrationtest.common.stub.backend.docume
 import org.hibernate.search.util.impl.integrationtest.common.stub.backend.index.StubDocumentWork;
 import org.hibernate.search.util.impl.integrationtest.common.stub.backend.index.StubIndexScopeWork;
 import org.hibernate.search.util.impl.integrationtest.common.stub.backend.search.StubSearchWork;
-import org.hibernate.search.util.impl.integrationtest.common.stub.backend.search.projection.impl.StubSearchProjection;
 
 import org.junit.rules.TestRule;
 import org.junit.runner.Description;
@@ -346,151 +339,6 @@ public class BackendMock implements TestRule {
 		private IndexScopeWorkCallListContext work(StubIndexScopeWork work, CompletableFuture<?> future) {
 			expectationConsumer.accept( new IndexScopeWorkCall( indexNames, work, future ) );
 			return this;
-		}
-	}
-
-	private class VerifyingStubBackendBehavior extends StubBackendBehavior {
-
-		private final Map<IndexFieldKey, CallBehavior> indexFieldAddBehaviors = new HashMap<>();
-
-		private final Map<String, CallQueue<PushSchemaCall>> pushSchemaCalls = new HashMap<>();
-
-		private final Map<String, CallQueue<DocumentWorkCall>> documentWorkCalls = new HashMap<>();
-
-		private final CallQueue<IndexScopeWorkCall> indexScopeWorkCalls = new CallQueue<>();
-
-		private final CallQueue<SearchWorkCall<?>> searchCalls = new CallQueue<>();
-
-		private final CallQueue<CountWorkCall> countCalls = new CallQueue<>();
-
-		void setIndexFieldAddBehavior(String indexName, String absoluteFieldPath, CallBehavior behavior) {
-			indexFieldAddBehaviors.put( new IndexFieldKey( indexName, absoluteFieldPath ), behavior );
-		}
-
-		CallQueue<PushSchemaCall> getPushSchemaCalls(String indexName) {
-			return pushSchemaCalls.computeIfAbsent( indexName, ignored -> new CallQueue<>() );
-		}
-
-		CallQueue<DocumentWorkCall> getDocumentWorkCalls(String indexName) {
-			return documentWorkCalls.computeIfAbsent( indexName, ignored -> new CallQueue<>() );
-		}
-
-		CallQueue<IndexScopeWorkCall> getIndexScopeWorkCalls() {
-			return indexScopeWorkCalls;
-		}
-
-		CallQueue<SearchWorkCall<?>> getSearchWorkCalls() {
-			return searchCalls;
-		}
-
-		CallQueue<CountWorkCall> getCountWorkCalls() {
-			return countCalls;
-		}
-
-		void resetExpectations() {
-			indexFieldAddBehaviors.clear();
-			pushSchemaCalls.clear();
-			documentWorkCalls.clear();
-			searchCalls.reset();
-		}
-
-		void verifyExpectationsMet() {
-			pushSchemaCalls.values().forEach( CallQueue::verifyExpectationsMet );
-			documentWorkCalls.values().forEach( CallQueue::verifyExpectationsMet );
-			searchCalls.verifyExpectationsMet();
-		}
-
-		@Override
-		public void onAddField(String indexName, String absoluteFieldPath) {
-			CallBehavior behavior = indexFieldAddBehaviors.get( new IndexFieldKey( indexName, absoluteFieldPath ) );
-			if ( behavior != null ) {
-				behavior.execute();
-			}
-		}
-
-		@Override
-		public void pushSchema(String indexName, StubIndexSchemaNode rootSchemaNode) {
-			getPushSchemaCalls( indexName )
-					.verify( new PushSchemaCall( indexName, rootSchemaNode ), PushSchemaCall::verify );
-		}
-
-		@Override
-		public void prepareDocumentWorks(String indexName, List<StubDocumentWork> works) {
-			CallQueue<DocumentWorkCall> callQueue = getDocumentWorkCalls( indexName );
-			works.stream()
-					.map( work -> new DocumentWorkCall( indexName, DocumentWorkCall.WorkPhase.PREPARE, work ) )
-					.forEach( call -> callQueue.verify( call, DocumentWorkCall::verify ) );
-		}
-
-		@Override
-		public CompletableFuture<?> executeDocumentWorks(String indexName, List<StubDocumentWork> works) {
-			CallQueue<DocumentWorkCall> callQueue = getDocumentWorkCalls( indexName );
-			return works.stream()
-					.map( work -> new DocumentWorkCall( indexName, DocumentWorkCall.WorkPhase.EXECUTE, work ) )
-					.<CompletableFuture<?>>map( call -> callQueue.verify( call, DocumentWorkCall::verify ) )
-					.reduce( (first, second) -> second )
-					.orElseGet( () -> CompletableFuture.completedFuture( null ) );
-		}
-
-		@Override
-		public CompletableFuture<?> prepareAndExecuteDocumentWork(String indexName, StubDocumentWork work) {
-			CallQueue<DocumentWorkCall> callQueue = getDocumentWorkCalls( indexName );
-			callQueue.verify(
-					new DocumentWorkCall( indexName, DocumentWorkCall.WorkPhase.PREPARE, work ),
-					DocumentWorkCall::verify
-			);
-			return callQueue.verify(
-					new DocumentWorkCall( indexName, DocumentWorkCall.WorkPhase.EXECUTE, work ),
-					DocumentWorkCall::verify
-			);
-		}
-
-		@Override
-		public <T> SearchResult<T> executeSearchWork(List<String> indexNames, StubSearchWork work,
-				FromDocumentFieldValueConvertContext convertContext,
-				LoadingContext<?, ?> loadingContext, StubSearchProjection<T> rootProjection) {
-			return searchCalls.verify(
-					new SearchWorkCall<>( indexNames, work, convertContext, loadingContext, rootProjection ),
-					(call1, call2) -> call1.verify( call2 )
-			);
-		}
-
-		@Override
-		public CompletableFuture<?> executeIndexScopeWork(List<String> indexNames, StubIndexScopeWork work) {
-			return indexScopeWorkCalls.verify(
-					new IndexScopeWorkCall( indexNames, work ),
-					IndexScopeWorkCall::verify
-			);
-		}
-
-		@Override
-		public long executeCountWork(List<String> indexNames) {
-			return countCalls.verify( new CountWorkCall( indexNames, null ), CountWorkCall::verify );
-		}
-	}
-
-	private static class IndexFieldKey {
-		final String indexName;
-		final String absoluteFieldPath;
-
-		private IndexFieldKey(String indexName, String absoluteFieldPath) {
-			this.indexName = indexName;
-			this.absoluteFieldPath = absoluteFieldPath;
-		}
-
-		@Override
-		public boolean equals(Object obj) {
-			if ( ! (obj instanceof IndexFieldKey) ) {
-				return false;
-			}
-			IndexFieldKey other = (IndexFieldKey) obj;
-			return Objects.equals( indexName, other.indexName )
-					&& Objects.equals( absoluteFieldPath, other.absoluteFieldPath );
-		}
-
-		@Override
-		public int hashCode() {
-			return Objects.hash( indexName, absoluteFieldPath );
 		}
 	}
